requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-storage
pandas
numpy
python-dotenv==0.19.1
#scikit-learn==1.2.2
psycopg2-binary==2.9.5
jwt==1.3.1
peewee==3.16.0
pydantic==1.10.7
passlib==1.7.4
python-jose==3.3.0
typing
bcrypt
cachetools
#tensorflow==1.5.0
joblib
gunicorn==20.1.0
python-multipart